# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "University management API - students, schedules, attendance",
        "title": "University API",
        "contact": {},
        "version": "1.0"
    },
    "host": "uni-server-29pn.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Email and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Email and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/attendance": {
            "get": {
                "tags": [
                    "attendance"
                ],
                "summary": "Get all attendance records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AttendanceRecord"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "attendance"
                ],
                "summary": "Create attendance record",
                "parameters": [
                    {
                        "description": "Attendance data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                                "$ref": "#/definitions/model.CreateAttendanceRequest"
                            }
                        }
                    ],
                    "responses": {
                        "201": {
                            "description": "Created",
                            "schema": {
                                "$ref": "#/definitions/model.AttendanceRecord"
                            }
                        }
                }
            }
        },
        "/faculties": {
            "get": {
                "tags": ["faculties"],
                "summary": "Get all faculties",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/model.FacultyResponse" }
                        }
                    }
                }
            },
            "post": {
                "consumes": ["application/json"],
                "tags": ["faculties"],
                "summary": "Create a faculty",
                "parameters": [
                    {
                        "description": "Faculty data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/model.CreateFacultyRequest" }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": { "$ref": "#/definitions/model.FacultyResponse" }
                    }
                }
            }
        },
        "/faculties/{id}": {
            "get": {
                "tags": ["faculties"],
                "summary": "Get faculty by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Faculty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/model.FacultyResponse" }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/groups": {
            "get": {
                "tags": ["groups"],
                "summary": "Get all groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/model.GroupResponse" }
                        }
                    }
                }
            },
            "post": {
                "consumes": ["application/json"],
                "tags": ["groups"],
                "summary": "Create a group",
                "parameters": [
                    {
                        "description": "Group data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/model.CreateGroupRequest" }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": { "$ref": "#/definitions/model.GroupResponse" }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "tags": ["groups"],
                "summary": "Get group by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/model.GroupResponse" }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/student/{id}": {
            "get": {
                "tags": [
                    "students"
                ],
                "summary": "Get student by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StudentResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students": {
            "get": {
                "tags": [
                    "students"
                ],
                "summary": "Get all students",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.StudentListResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Create a student",
                "parameters": [
                    {
                        "description": "Student data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateStudentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.StudentResponse"
                        }
                    }
                }
            }
        },
        "/subjects": {
            "get": {
                "tags": ["subjects"],
                "summary": "Get all subjects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/model.SubjectResponse" }
                        }
                    }
                }
            },
            "post": {
                "consumes": ["application/json"],
                "tags": ["subjects"],
                "summary": "Create a subject",
                "parameters": [
                    {
                        "description": "Subject data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": { "$ref": "#/definitions/model.CreateSubjectRequest" }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": { "$ref": "#/definitions/model.SubjectResponse" }
                    }
                }
            }
        },
        "/subjects/{id}": {
            "get": {
                "tags": ["subjects"],
                "summary": "Get subject by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subject ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/model.SubjectResponse" }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AttendanceRecord": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "student_id": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "integer"
                },
                "visit_day": {
                    "type": "string"
                },
                "visited": {
                    "type": "boolean"
                }
            }
        },
        "model.CreateFacultyRequest": {
            "type": "object",
            "properties": {
                "name": { "type": "string" }
            }
        },
...
        "model.FacultyResponse": {
            "type": "object",
            "properties": {
                "id": { "type": "integer" },
                "name": { "type": "string" }
            }
        },
        "model.GroupResponse": {
            "type": "object",
            "properties": {
                "id": { "type": "integer" },
                "name": { "type": "string" },
                "faculty_id": { "type": "integer" },
                "faculty_name": { "type": "string" }
            }
        },
        "model.AuthRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.CreateAttendanceRequest": {
            "type": "object",
            "required": ["student_id", "subject_id", "visit_day", "visited"],
            "properties": {
                "student_id": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "integer"
                },
                "visit_day": {
                    "type": "string"
                },
                "visited": {
                    "type": "boolean"
                }
            }
        },
...
        "model.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
...
    }
}